Move action and action creator from actions/singleEventActions to actions/eventActions
