+++ client/actions/eventActions.js
@@ -16,3 +16,18 @@ export function fetchEvents() {
       });
   };
 }
+export function fetchSingleEventSuccess(event) {
+  return { type: types.FETCH_SINGLE_EVENT_SUCCESS, event };
+}
+
+export function fetchSingleEvent(eventId) {
+  return function (dispatch) {
+    return EventApi.getOne(eventId)
+      .then((response) => {
+        dispatch(fetchSingleEventSuccess(response.data.event));
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  };
+}
